Parse body as JSON regardless of Content-Type
The node AWS SDK for Lambda sends requests with a "Content-Type: application/x-amz-json-1.0", according to my reading of the source. This allows support for that, and use of curl and other manual API use. We might instead want to use a function for `type` parameter, and have it check the content-type against 'application/json' and the aforementioned amz content type.